feat: Add Cart module and core infrastructure re-architecting
A new **Cart** microservice has been added, complete with its API, Application, Controllers, Domain, and Infrastructure layers.

Significant architectural changes have been introduced across the `Common` projects and for system orchestration:

**New Infrastructure & Architecture:**

* **Aspire Integration:** Introduced the `Ecommerce.AppHost` project using **Aspire** for service orchestration, Redis caching, and RabbitMQ messaging setup.
* **Common API Service Defaults:** Added `Common.Api.ServiceDefaults` for standardized health checks and OpenTelemetry configuration.
* **Redis & Messaging Support:** Added **Redis** cache management services and **RabbitMQ** messaging infrastructure to `Common.Infrastructure` for distributed communication and state sharing (e.g., seeding dependencies).
* **Seeder Rework:** The seeding mechanism (`Common.Infrastructure.Seeder`) was refactored to be asynchronous, dependency-aware (using Redis for coordination), and decoupled from the main database context.

**Domain & DTO Changes:**

* **Value Objects:** Introduced `Quantity` and `Price` value objects (with associated DTOs and database configurations) to `Common.Domain` for improved domain modeling. Added `Unit` enum.
* **Security Domain Migration:** All Security-related entities (`User`, `Persona`), DTOs, Controllers, and related infrastructure components (AuthServices, Configurators, Seeders) were moved from `Common` to the new dedicated **Security** module.
* **Module Stubs Added:** Created foundational projects (Domain, Application, Controllers, API) for **Inventory, Product, Order, Payment, Invoice, Notification, and Shipping** services to complete the microservice structure.

These changes establish the foundation for a scalable, observable, and modular microservice architecture.

+++ Common/Domain/Common.Domain/ValueObjects/Coordinates.cs
@@ -3,8 +3,8 @@ namespace Common.Domain.ValueObjects
 {
     public class Coordinates
     {
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
 
         public override bool Equals(object? obj)
         {
